services/data-capture/pyproject.toml
[tool.poetry]
name = "data-capture"
version = "0.1.0"
description = "Data capture service for KindPath"
authors = ["KindPath Foundation"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
kafka-python = "^2.0.2"
boto3 = "^1.34.0"
cryptography = "^41.0.7"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"